requirements.txt
apache-airflow==3.1.0
beautifulsoup4==4.14.2
boto3==1.35.21
pandas==2.3.3
pendulum==3.1.0
Requests==2.32.5
